:sparkles: Feature #27: 로그인 안 했을 경우 ui 구현
